requirements.txt
pyfiglet>=1.0.2
colorama>=0.4.6
rich>=13.7.0
click>=8.1.7
Pillow>=10.2.0
jinja2>=3.1.3
PyYAML>=6.0.1